requirements.txt
click==8.0.0
dlib==19.22.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.0
Flask-JSGlue==0.3.1
Flask-MySQL==1.5.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
opencv-contrib-python==4.1.0.25
paho-mqtt==1.5.1
Pillow==8.2.0
PyMySQL==1.0.2
RPi.GPIO==0.7.0
#pip install --extra-index-url https://google-coral.github.io/py-repo/ tflite_runtime
Werkzeug==2.0.1